+++ library/src/methods/getDefaults/getDefaults.ts
@@ -20,7 +20,7 @@ export function getDefaults<TSchema extends SchemaWithMaybeDefault>(
   schema: TSchema
 ): DefaultValues<TSchema> | undefined {
   // If schema contains a default function, set its default value
-  if (schema.default) {
+  if (schema.default !== undefined) {
     return getDefault(schema);
   }
   // Otherwise, check if schema is of kind object or tuple