+++ library/src/methods/getDefaults/getDefaultsAsync.ts
@@ -21,7 +21,7 @@ export async function getDefaultsAsync<
   TSchema extends SchemaWithMaybeDefault | SchemaWithMaybeDefaultAsync
 >(schema: TSchema): Promise<DefaultValues<TSchema> | undefined> {
   // If schema contains a default function, set its default value
-  if (schema.default) {
+  if (schema.default !== undefined) {
     return getDefaultAsync(schema);
   }
   // Otherwise, check if schema is of kind object or tuple